requirements.txt
autogen-agentchat==0.7.2
autogen-ext[openai]==0.7.2
dotenv
pypinyin
pymongo==4.6.1
fastapi==0.111.0
uvicorn==0.29.0
pydantic>=2.10.0
pytest==7.4.4
requests==2.31.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
cryptography>=42.0.0
PyJWT>=2.8.0
pymysql>=1.1.0
